requirements.txt
# FastAPI & Server
fastapi==0.109.0

# VPS Provisioning
boto3>=1.34.0
stripe>=7.0.0
uvicorn[standard]==0.27.0
python-multipart>=0.0.9

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
pgvector==0.2.4
alembic==1.13.1

# For local development with SQLite
aiosqlite==0.19.0
sqlite-vss==0.1.2

# Embeddings & NLP
sentence-transformers==2.3.1
spacy>=3.7.5
numpy==1.26.3
openai>=1.0.0
tiktoken>=0.5.0  # Token counting for chat

# Document Processing
pypdf==4.0.1
python-docx==1.1.0
markdown==3.5.2
python-magic==0.4.27
chardet==5.2.0

# Media Processing
pillow==10.2.0
# moviepy==1.0.3  # Optional: for video processing
# openai-whisper==20231117  # Optional: for audio transcription

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]>=2.7.2,<3.0.0
pydantic-settings>=2.5.2

# Utils
python-dateutil==2.8.2
httpx>=0.28.1
tenacity==8.2.3
apscheduler==3.10.4  # Scheduled tasks for decay/consolidation

# Telegram Bot
python-telegram-bot>=20.0

# Discord
discord.py>=2.3.0

# Slack
slack-bolt>=1.18.0
slack-sdk>=3.27.0

# Cohere (Re-ranker)
cohere>=5.0.0

# Anthropic
anthropic>=0.30.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# MCP Server
fastmcp==2.9.2

# Dev
black==24.1.0
ruff==0.1.14
